refactor: unify isSystemLib into internal/funkutil
Two hand-maintained regexes served the same purpose (skip bulky
system/runtime libraries with thousands of symbols that aren't worth
wildcard-tracing) with different, never-reconciled lists: cmd/enumerate.go
(install-time ldd enumeration) was missing libthread_db/libglib/
libgobject/libgthread/libgio/libcap/libattr/libpcre/libselinux/
libmount/libblkid/libuuid/libpam/libaudit/libdbus/libsystemd/libudev;
cmd/shim_binary/tracer.go (runtime dlopen discovery) was missing
libresolv/libnsl/libutil/libcrypt/libanl. Merged into one shared
IsSystemLib in funkutil using the union of both lists, verified against
the union of both existing test suites (now one consolidated test).

ponytail-audit finding #4 (tackled before #3 since it shares the same
duplicate-logic shape as the funcIsRelevant unification).
